+++ tests/Core/Migration/Migration1761903189AdjustSwagMigrationLoggingTableTest.php
@@ -8,11 +8,16 @@
 namespace Core\Migration;
 
 use PHPUnit\Framework\TestCase;
+use Shopware\Core\Framework\Log\Package;
 use Shopware\Core\Framework\Test\TestCaseBase\KernelLifecycleManager;
 use Shopware\Core\Framework\Test\TestCaseBase\KernelTestBehaviour;
 use SwagMigrationAssistant\Core\Migration\Migration1761903189AdjustSwagMigrationLoggingTable;
-use SwagMigrationAssistant\Core\Migration\TableHelperTrait;
+use SwagMigrationAssistant\Test\TableHelperTrait;
 
+/**
+ * @internal
+ */
+#[Package('after-sales')]
 class Migration1761903189AdjustSwagMigrationLoggingTableTest extends TestCase
 {
     use KernelTestBehaviour;
@@ -21,20 +26,15 @@ class Migration1761903189AdjustSwagMigrationLoggingTableTest extends TestCase
     public function testUpdate(): void
     {
         $connection = KernelLifecycleManager::getConnection();
-        if ($this->indexExists($connection, 'swag_migration_logging', 'idx.entity_id')) {
-            $connection->executeStatement('ALTER TABLE `swag_migration_logging` DROP INDEX `idx.entity_id`;');
-            static::assertFalse($this->indexExists($connection, 'swag_migration_logging', 'idx.entity_id'));
-        }
 
-        if ($this->columnExists($connection, 'swag_migration_logging', 'entity_id')) {
-            $connection->executeStatement('ALTER TABLE `swag_migration_logging` DROP COLUMN `entity_id`;');
-            static::assertFalse($this->columnExists($connection, 'swag_migration_logging', 'entity_id'));
-        }
+        $this->dropIndex($connection, 'swag_migration_logging', 'idx.entity_id');
+        static::assertFalse($this->indexExists($connection, 'swag_migration_logging', 'idx.entity_id'));
 
-        if (!$this->columnExists($connection, 'swag_migration_logging', 'used_mapping')) {
-            $connection->executeStatement('ALTER TABLE `swag_migration_logging` ADD COLUMN `used_mapping` JSON;');
-            static::assertTrue($this->columnExists($connection, 'swag_migration_logging', 'used_mapping'));
-        }
+        $this->dropColumn($connection, 'swag_migration_logging', 'entity_id');
+        static::assertFalse($this->columnExists($connection, 'swag_migration_logging', 'entity_id'));
+
+        $this->addColumn($connection, 'swag_migration_logging', 'used_mapping', 'JSON');
+        static::assertTrue($this->columnExists($connection, 'swag_migration_logging', 'used_mapping'));
 
         $migration = new Migration1761903189AdjustSwagMigrationLoggingTable();
         $migration->update($connection);
